refactor(draft-packages): rename scss files to use .module.scss
- guidance-block
- LoadingSkeleton stories updated to not use guidance-block styles
