+++ draft-packages/guidance-block/KaizenDraft/GuidanceBlock/GuidanceBlock.module.scss
@@ -5,7 +5,6 @@
 @import "~@kaizen/design-tokens/sass/animation";
 @import "~@kaizen/design-tokens/sass/spacing";
 @import "~@kaizen/component-library/styles/responsive";
-@import "./GuidanceBlockBreakpoints.scss";
 @import "./mixins";
 
 $bannerPadding: $spacing-lg;
