+++ draft-packages/guidance-block/KaizenDraft/GuidanceBlock/_mixins.scss
@@ -8,3 +8,9 @@
   padding: 0;
   transition: none; // override Murmur global styles :(
 }
+
+@mixin guidance-block-under-480 {
+  @media (max-width: 480px) {
+    @content;
+  }
+}
